fixed error with merge
